feat: 新しいexportDictionary APIを利用することでユーザ辞書の書き出しロジックを変換機側に統合
